requirements.txt
uvicorn==0.15.0
pydantic==1.8.2
fastapi==0.68.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.6.5
